Switch to extension-ci-tools based CI (#15)
With this PR, duckdb-prql should be good to go for going into the community extensions repo.

I switched a few things around to stay in line with the most recent extension template. This simplifies things a little bit and should ensure that when CI passes in this repo, it will pass in the the community extensions repo.